add folder files
