+++ starwars/swapi/src/App.css
@@ -2,7 +2,7 @@
 @import url('https://fonts.googleapis.com/css?family=Roboto+Mono');
 
 .Background {
-  background-image: url('./img/starwars.jpeg');
+  background-image: url('./assets/grogu.png');
   margin: auto;
   width: auto;
   height: 999px;
