+++ starwars/swapi/src/assets/starwars.jpeg
